add create, sum and contains
