+++ src/IntrinsicsGeneric.UnitTests/Simd/Avx.cs
@@ -1,5 +1,6 @@
 ﻿using System.Runtime.Intrinsics;
 using IntrinsicsGeneric.Extensions;
+using IntrinsicsGeneric.Helpers;
 using IntrinsicsGeneric.Simd;
 using Xunit;
