+++ src/IntrinsicsGeneric.UnitTests/Simd/Avx2.cs
@@ -1,5 +1,6 @@
 ﻿using System.Runtime.Intrinsics;
 using IntrinsicsGeneric.Extensions;
+using IntrinsicsGeneric.Helpers;
 using IntrinsicsGeneric.Simd;
 using Xunit;
 using AVX2 = System.Runtime.Intrinsics.X86.Avx2;
