+++ src/IntrinsicsGeneric.UnitTests/Simd/Sse2.cs
@@ -1,5 +1,6 @@
 ﻿using System.Runtime.Intrinsics;
 using IntrinsicsGeneric.Extensions;
+using IntrinsicsGeneric.Helpers;
 using IntrinsicsGeneric.Simd;
 using Xunit;
 using SSE2 = System.Runtime.Intrinsics.X86.Sse2;
