+++ src/IntrinsicsGeneric.UnitTests/VectorHelperTests/Contains.cs
@@ -1,62 +1,19 @@
-using IntrinsicsGeneric.Extensions;
+using IntrinsicsGeneric.Helpers;
 using Xunit;
 
-namespace IntrinsicsGeneric.UnitTests.Extensions
+namespace IntrinsicsGeneric.UnitTests.VectorHelperTests
 {
-    public class VectorExtensionsTest
+    public class Contains
     {
-        [Theory]
-        [InlineData(1, 4)]
-        [InlineData(1U, 4U)]
-        [InlineData(1F, 4F)]
-        [InlineData(1L, 2L)]
-        [InlineData(1UL, 2UL)]
-        [InlineData(1D, 2D)]
-        [InlineData((short) 1, (short) 8)]
-        [InlineData((ushort) 1, (ushort) 8)]
-        [InlineData((byte) 1, (byte) 16)]
-        [InlineData((sbyte) 1, (sbyte) 16)]
-        public void Vector128Sum<T>(T value, T expected)
-            where T : unmanaged
-        {
-            var v = VectorHelper<T>.CreateVector128(value);
-
-            var result = v.Sum();
-
-            Assert.Equal(expected, result);
-        }
-
-        [Theory]
-        [InlineData(1, 8)]
-        [InlineData(1U, 8U)]
-        [InlineData(1F, 8F)]
-        [InlineData(1L, 4L)]
-        [InlineData(1UL, 4UL)]
-        [InlineData(1D, 4D)]
-        [InlineData((short) 1, (short) 16)]
-        [InlineData((ushort) 1, (ushort) 16)]
-        [InlineData((byte) 1, (byte) 32)]
-        [InlineData((sbyte) 1, (sbyte) 32)]
-        public void Vector256Sum<T>(T value, T expected)
-            where T : unmanaged
-        {
-            var v = VectorHelper<T>.CreateVector256(value);
-
-            var result = v.Sum();
-
-            Assert.Equal(expected, result);
-        }
-        
-                #region Contains
-
         [Fact]
-        public void Contains_True()
+        public void ShouldContainValueOfVector()
         {
-            /*Assert.True(VectorHelper<byte>.Contains(
+            Assert.True(VectorHelper<byte>.Contains(new byte[]
+            {
                 15, 75, 42, 37, 79, 67, 7, 120, 15, 75, 42, 37, 79, 67, 7, 120,
-                15, 75, 42, 37, 79, 67, 7, 120, 15, 75, 42, 137, 79, 67, 7, 120
-            ), 137);
-            */
+                15, 75, 1, 37, 79, 67, 7, 120, 15, 75, 42, 1, 79, 67, 7, 120
+            }, 37));
+            
             
             Assert.True(VectorHelper<sbyte>.Contains(new sbyte[]
             {
@@ -78,7 +35,7 @@ namespace IntrinsicsGeneric.UnitTests.Extensions
         }
         
         [Fact]
-        public void Contains_False()
+        public void ShouldNotContainValueOfVector()
         {
             Assert.False(VectorHelper<byte>.Contains(new byte[]
             {
@@ -103,7 +60,5 @@ namespace IntrinsicsGeneric.UnitTests.Extensions
             Assert.False(VectorHelper<float>.Contains(new float[] { 15, 75, 42, 37, 79, 67.9f, 7, 120 }, 67.19f));
             Assert.False(VectorHelper<double>.Contains(new double[] { 42, 37, 42, 37.98 }, 37.9));
         }
-        
-        #endregion
     }
 }
