+++ src/IntrinsicsGeneric.sln.DotSettings.user
@@ -12,5 +12,7 @@
     &lt;TestId&gt;xUnit::C4EAD125-79E9-4C24-9326-4B36832098BA::net5.0::IntrinsicsGeneric.UnitTests.Extensions.VectorHelper.Contains_False&lt;/TestId&gt;&#xD;
     &lt;TestId&gt;xUnit::C4EAD125-79E9-4C24-9326-4B36832098BA::net5.0::IntrinsicsGeneric.UnitTests.Extensions.VectorHelper.Contains_True&lt;/TestId&gt;&#xD;
     &lt;TestId&gt;xUnit::C4EAD125-79E9-4C24-9326-4B36832098BA::net5.0::IntrinsicsGeneric.UnitTests.Extensions.VectorHelper&lt;/TestId&gt;&#xD;
+    &lt;TestId&gt;xUnit::C4EAD125-79E9-4C24-9326-4B36832098BA::net5.0::IntrinsicsGeneric.UnitTests.ExtensionsTests.Contains.ShouldContainValueOfVector&lt;/TestId&gt;&#xD;
+    &lt;TestId&gt;xUnit::C4EAD125-79E9-4C24-9326-4B36832098BA::net5.0::IntrinsicsGeneric.UnitTests.ExtensionsTests.Contains.ShouldNotContainValueOfVector&lt;/TestId&gt;&#xD;
   &lt;/TestAncestor&gt;&#xD;
 &lt;/SessionState&gt;</s:String></wpf:ResourceDictionary>
