+++ src/IntrinsicsGeneric/Extensions/VectorExtensions.cs
@@ -1,63 +1,21 @@
-﻿using IntrinsicsGeneric.InternalHelpers;
-using System;
+﻿using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Runtime.CompilerServices;
 using System.Runtime.Intrinsics;
 using System.Numerics;
-using System.Runtime.Intrinsics.X86;
-using IntrinsicsGeneric.Simd;
 
 namespace IntrinsicsGeneric.Extensions
 {
-    public static class VectorExtensions
+    public static partial class VectorExtensions
     {
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static T Sum<T>(this Vector128<T> vector)
-            where T : unmanaged
-        {
-            var sum = default(T);
-            for (var i = 0; i < Vector128<T>.Count; i++)
-            {
-                sum = MathUnsafe<T>.Add(sum, vector.GetElement(i));
-            }
-
-            return sum;
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static T Sum<T>(this Vector256<T> a)
-            where T : unmanaged
-        {
-            var sum = default(T);
-            for (var i = 0; i < Vector256<T>.Count; i++)
-            {
-                sum = MathUnsafe<T>.Add(sum, a.GetElement(i));
-            }
-            return sum;
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static bool Contains<T>(this Vector128<T> vector, T value)
-            where T : unmanaged
-        {
-            var element = VectorHelper<T>.CreateVector128(value);
-            var mask = Sse41<T>.CompareEqual(vector, element);
-            return !Sse41<T>.TestAllZeros(mask, mask);
-        }
-        
-        public static bool Contains<T>(this Vector256<T> vector, T value)
-            where T : unmanaged
-        {
-            var element = VectorHelper<T>.CreateVector256(value);
-            var mask = Avx2<T>.CompareEqual(vector, element);
-            return !Avx<T>.TestAllZeros(mask, mask);
-        }
-
+        /// <summary>
+        /// Converts Vector128<T/> to BitArray.
+        /// </summary>
         public static unsafe BitArray ToBitArray<T>(this Vector128<T> vector)
             where T : unmanaged
         {
-            List<byte> bytes = new List<byte>(sizeof(T) * Vector128<T>.Count);
+            var bytes = new List<byte>(sizeof(T) * Vector128<T>.Count);
 
             for (int i = 0; i < Vector128<T>.Count; i++)
             {
@@ -67,10 +25,13 @@ namespace IntrinsicsGeneric.Extensions
             return new BitArray(bytes.ToArray());
         }
 
+        /// <summary>
+        /// Converts Vector256<T/> to BitArray.
+        /// </summary>
         public static unsafe BitArray ToBitArray<T>(this Vector256<T> vector)
             where T : unmanaged
         {
-            List<byte> bytes = new List<byte>(sizeof(T) * Vector256<T>.Count);
+            var bytes = new List<byte>(sizeof(T) * Vector256<T>.Count);
 
             for (int i = 0; i < Vector256<T>.Count; i++)
             {
@@ -132,12 +93,18 @@ namespace IntrinsicsGeneric.Extensions
             throw new NotSupportedException();
         }
         
+        /// <summary>
+        /// Converts Vector<T/> to Vector128<T/>.
+        /// </summary>
         public static Vector128<T> AsVector128<T>(this Vector<T> value)
             where T : struct
         {
             return Unsafe.As<Vector<T>, Vector128<T>>(ref value);
         }
 
+        /// <summary>
+        /// Converts Vector<T/> to Vector256<T/>.
+        /// </summary>
         public static Vector256<T> AsVector256<T>(this Vector<T> value)
             where T : struct
         {
